int to byte for request body bytes

+++ src/main/java/common/streaming/BodyBytesDequeue.java
@@ -1,5 +1,5 @@
 package common.streaming;
 
 public interface BodyBytesDequeue {
-    int dequeue();
+    byte dequeue();
 }
